Logger component cont'd - search edit
