+++ src/layouts/MainLayout.vue
@@ -107,48 +107,43 @@
           after-class="bg-white logger-class"
         >
 
+          <!--Router View-->
           <template v-slot:before>
             <router-view />
           </template>
 
+          <!--Logger Component-->
           <template v-slot:after>
             <q-page-sticky expand position="top" style="position: sticky; transform: none">
-              <q-toolbar class="bg-grey-4 shadow-1">
+              <q-toolbar class="bg-grey-4">
                 <q-toolbar-title class="text-grey-8 text-subtitle1">
                   <q-icon name="fas fa-columns fa-sm" />
                   Console
                 </q-toolbar-title>
-
                 <q-space />
-
-                <template v-if="splitterModel < limits[1]">
-                  <div v-if="searchKey" class="text-caption">
-                    {{ matchCount }} {{ matchCount > 1 ? ' matches' : ' match' }}
-                  </div>
-                  <q-input dense
-                           standout="bg-grey-8"
-                           placeholder="Search..."
-                           v-model="searchKey"
-                           class="q-ml-md"
-                  >
-                    <template v-slot:append>
-                      <q-icon v-if="searchKey === ''" name="search" />
-                      <q-icon v-else name="clear" class="cursor-pointer" @click="searchKey = ''" />
-                    </template>
-                  </q-input>
-                </template>
-                <q-btn
-                  round
-                  flat
-                  dense
-                  color="grey-8"
-                  :icon="splitterModel >= limits[1] ? 'keyboard_arrow_up' : 'keyboard_arrow_down'"
-                  @click="splitterModel >= limits[1] ? splitterModel = 70 : splitterModel = limits[1]"
-                />
+                <div class="row items-center q-gutter-xs">
+                  <template v-if="splitterModel < limits[1]">
+                    <div v-if="searchKey" class="text-caption">
+                      {{ matchCount }} {{ matchCount > 1 ? ' matches' : ' match' }}
+                    </div>
+                    <q-input dense rounded standout="bg-grey-5" placeholder="Search..." v-model="searchKey"
+                             class="q-ml-md" input-class="text-grey-9" @keydown.esc="searchKey = ''"
+                    >
+                      <template v-slot:append>
+                        <q-icon v-if="searchKey === ''" name="search" color="grey-8" />
+                        <q-icon v-else name="clear" color="grey-8" class="cursor-pointer" @click="searchKey = ''" />
+                      </template>
+                    </q-input>
+                  </template>
+                  <q-btn round flat dense color="grey-8"
+                         :icon="splitterModel >= limits[1] ? 'keyboard_arrow_up' : 'keyboard_arrow_down'"
+                         @click="splitterModel >= limits[1] ? splitterModel = 70 : splitterModel = limits[1]"
+                  />
+                </div>
               </q-toolbar>
             </q-page-sticky>
             <template v-if="splitterModel < limits[1]">
-              <div class="q-mt-xs q-mx-sm q-mb-md text-caption" v-html="highlight(logger, searchKey)" />
+              <Logger v-model="matchCount" :searchKey="searchKey" />
             </template>
           </template>
 
@@ -161,7 +156,11 @@
 <script lang="ts">
   import { Component, Vue, Watch } from 'vue-property-decorator'
 
-  @Component
+  @Component({
+    components: {
+      Logger: () => import('@/components/Logger.vue')
+    }
+  })
   export default class MainLayout extends Vue {
     private splitterModel: any = this.limits[1]
     private miniState: boolean = true
@@ -189,21 +188,6 @@
           logDiv.scrollTo(0, logDiv.scrollHeight)
         }, 0)
       }
-    }
-
-    highlight (text: string, search: string): string {
-      let count = 0
-      if (!search)
-        return text
-
-      search = search.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, '\\$&')
-      const regex = new RegExp(`(${search})(?!([^<]+)?>)`, 'gi')
-      const filtered = text.replace(regex, match => {
-        count += 1
-        return '<span style="background:#9CC3F6">' + match + '</span>'
-      })
-      this.matchCount = count
-      return filtered
     }
   }
 </script>
